# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  name: string;
  /** coffee | bites */
  category: string;
  price: number;
  description: string;
  /**
     * Optional highlight tag e.g. "Signature", "Chef's Pick"
     * @nullable
     */
  tag?: string | null;
  isSpicy?: boolean;
}

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  date: string;
  timeSlot: string;
  players: number;
  /** @nullable */
  notes?: string | null;
  /** confirmed | cancelled */
  status: string;
  createdAt: string;
}

export interface BookingInput {
  name: string;
  email: string;
  phone: string;
  /** YYYY-MM-DD */
  date: string;
  /** e.g. 09:00 - 10:00 */
  timeSlot: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  players: number;
  notes?: string;
}

export interface TimeSlot {
  slot: string;
  available: boolean;
}

export interface ContactInquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** private | corporate | tournament | general */
  inquiryType: string;
  message: string;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  phone: string;
  inquiryType: string;
  message: string;
}

export interface Stats {
  totalBookings: number;
  todayBookings: number;
  /** @nullable */
  popularSlot: string | null;
  /** @nullable */
  popularPlayers: number | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListMenuItemsParams = {
category?: string;
};

export type GetAvailableSlotsParams = {
/**
 * Date in YYYY-MM-DD format
 */
date: string;
};