# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  name: string;
  email: string;
  phone: string;
  date: string;
  timeSlot: string;
  players: number;
  /** @nullable */
  notes?: string | null;
  /** confirmed | cancelled */
  status: string;
  createdAt: string;
}