# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  name: string;
  email: string;
  phone: string;
  /** YYYY-MM-DD */
  date: string;
  /** e.g. 09:00 - 10:00 */
  timeSlot: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  players: number;
  notes?: string;
}