# lib/api-zod/src/generated/types/contactInquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** private | corporate | tournament | general */
  inquiryType: string;
  message: string;
  createdAt: string;
}