# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  /** coffee | bites */
  category: string;
  price: number;
  description: string;
  /**
     * Optional highlight tag e.g. "Signature", "Chef's Pick"
     * @nullable
     */
  tag?: string | null;
  isSpicy?: boolean;
}