# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalBookings: number;
  todayBookings: number;
  /** @nullable */
  popularSlot: string | null;
  /** @nullable */
  popularPlayers: number | null;
}