# lib/api-zod/src/generated/types/timeSlot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SCAVA Café & Pickleball API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSlot {
  slot: string;
  available: boolean;
}